[S12P31B106-308]refactor:EntityType

+++ backend/src/main/java/com/voiceprint/backend/api/chat/dto/SessionStartRequestDTO.java
@@ -10,5 +10,5 @@ import lombok.Setter;
 @NoArgsConstructor
 public class SessionStartRequestDTO {
     @NotNull
-    private Long chatbotId;
+    private Byte chatbotId;
 }
