+++ backend/src/main/java/com/voiceprint/backend/domain/Repository/GroupInviteRepository.java
@@ -6,11 +6,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDateTime;
 import java.util.Optional;
 
-public interface GroupInviteRepository extends JpaRepository<GroupInvite, Long> {
-    Optional<GroupInvite> findFirstByGroupIdAndExpiredAtAfter(Long groupId, LocalDateTime now);
+public interface GroupInviteRepository extends JpaRepository<GroupInvite, Integer> {
+    Optional<GroupInvite> findFirstByGroupIdAndExpiredAtAfter(Integer groupId, LocalDateTime now);
 
     // 최신 그룹 초대 엔티티 조회
-    Optional<GroupInvite> findTopByGroupIdOrderByCreatedAtDesc(Long groupId);
+    Optional<GroupInvite> findTopByGroupIdOrderByCreatedAtDesc(Integer groupId);
 
     Optional<GroupInvite> findByInviteCode(String code);
 }
