+++ backend/src/main/java/com/voiceprint/backend/service/alarm/DiaryReminderScheduler.java
@@ -50,7 +50,7 @@ public class DiaryReminderScheduler {
                 continue;  // 알람 시간 없음, 알람 꺼짐, 알람 시각 아님 → 패스
             }
 
-            Long userId = user.getId();
+            Integer userId = user.getId();
             String sessionKey = session_key + ":" +userId;
 
             Object statusObj = redisTemplate.opsForHash().get(sessionKey,"status");
