+++ backend/src/main/java/com/voiceprint/backend/service/alarm/SseService.java
@@ -16,17 +16,17 @@ public class SseService {
     private final SseEmitterManager emitterManager;
 
     // 클라이언트가 구독 요청할 때 호출
-    public SseEmitter subscribe(Long userId) {
+    public SseEmitter subscribe(Integer userId) {
         return emitterManager.add(userId); // Emitter 생성 및 저장
     }
 
     // 실제 알림을 전송하는 메서드 : 테스트용
-    public void sendNotification(Long userId, String eventName, NotificationDTO payload) {
+    public void sendNotification(Integer userId, String eventName, NotificationDTO payload) {
         emitterManager.sendTo(userId, eventName, payload);
     }
 
     // 구독중인 사용자 조회하는  : 그룹 용
-    public Set<Long> getSubscribedUserIds() {
+    public Set<Integer> getSubscribedUserIds() {
         return emitterManager.getSubscribedUserIds();
     }
 }
