+++ backend/src/main/java/com/voiceprint/backend/service/comment/CommentService.java
@@ -40,7 +40,7 @@ public class CommentService {
 
     @Transactional
     // 댓글 작성
-    public CommentCreateResponseDTO saveComment (long userId, long groupDiaryId, CommentCreatRequestDTO commentCreatRequestDTO) {
+    public CommentCreateResponseDTO saveComment (Integer userId, Integer groupDiaryId, CommentCreatRequestDTO commentCreatRequestDTO) {
 
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new UserNotFoundException("User를 찾을 수 없습니다."));
@@ -151,7 +151,7 @@ public class CommentService {
 
     // 댓글 삭제
     @Transactional
-    public void deleteComment (int commentId, long userId) {
+    public void deleteComment (Integer commentId, Integer userId) {
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new CommentNotFoundException("댓글이 없습니다."));
